+++ packages/sources/svelte/src/components/index.js
@@ -1,5 +1,7 @@
 // Actions
 export { default as VtmnButton } from './actions/VtmnButton/VtmnButton.svelte';
+export { default as VtmnDropdown } from './actions/VtmnDropdown/VtmnDropdown.svelte';
+export { default as VtmnDropdownItem } from './actions/VtmnDropdownItem/VtmnDropdownItem.svelte';
 export { default as VtmnLink } from './actions/VtmnLink/VtmnLink.svelte';
 
 // Forms
@@ -15,6 +17,8 @@ export { default as VtmnRating } from './indicators/VtmnRating/VtmnRating.svelte
 export { default as VtmnTag } from './indicators/VtmnTag/VtmnTag.svelte';
 
 // Navigation
+export { default as VtmnBreadcrumb } from './navigation/VtmnBreadcrumb/VtmnBreadcrumb.svelte';
+export { default as VtmnBreadcrumbItem } from './navigation/VtmnBreadcrumbItem/VtmnBreadcrumbItem.svelte';
 export { default as VtmnSearch } from './navigation/VtmnSearch/VtmnSearch.svelte';
 export { default as VtmnNavbar } from './navigation/VtmnNavbar/VtmnNavbar.svelte';
 export { default as VtmnTabs } from './navigation/VtmnTabs/VtmnTabs.svelte';