+++ packages/sources/svelte/src/components/overlays/VtmnModal/test/VtmnModalWithSlots.svelte
@@ -1,8 +1,0 @@
-<script>
-  import VtmnModal from '../VtmnModal.svelte';
-</script>
-
-<VtmnModal {...$$restProps}>
-  <p slot="description">description</p>
-  <div slot="actions">actions</div>
-</VtmnModal>