+++ packages/sources/vue/src/components/index.ts
@@ -6,11 +6,16 @@ import VtmnBadge from './indicators/VtmnBadge';
 import VtmnLoader from './indicators/VtmnLoader';
 import VtmnPrice from './indicators/VtmnPrice';
 import VtmnTag from './indicators/VtmnTag';
+import VtmnSearch from './navigation/VtmnSearch';
 import VtmnAlert from './overlays/VtmnAlert';
+import VtmnModal from './overlays/VtmnModal';
 import VtmnPopover from './overlays/VtmnPopover';
+import VtmnSnackbar from './overlays/VtmnSnackbar';
+import VtmnToast from './overlays/VtmnToast';
 import VtmnTooltip from './overlays/VtmnTooltip';
 import VtmnCheckbox from './selection-controls/VtmnCheckbox';
 import VtmnChip from './selection-controls/VtmnChip';
+import VtmnQuantity from './selection-controls/VtmnQuantity';
 import VtmnRadioButton from './selection-controls/VtmnRadioButton';
 import VtmnToggle from './selection-controls/VtmnToggle';
 import VtmnAccordion from './structure/VtmnAccordion';
@@ -25,13 +30,18 @@ export {
   VtmnBadge,
   VtmnLoader,
   VtmnPrice,
+  VtmnTag,
+  VtmnSearch,
+  VtmnAlert,
+  VtmnModal,
+  VtmnPopover,
+  VtmnSnackbar,
+  VtmnToast,
+  VtmnTooltip,
   VtmnCheckbox,
   VtmnChip,
+  VtmnQuantity,
   VtmnRadioButton,
-  VtmnTag,
-  VtmnAlert,
-  VtmnPopover,
-  VtmnTooltip,
   VtmnToggle,
   VtmnAccordion,
   VtmnDivider,