+++ packages/sources/vue/src/components/navigation/VtmnSearch/index.ts
@@ -1,0 +1,8 @@
+import { App as Application } from 'vue';
+import VtmnSearch from './VtmnSearch.vue';
+
+VtmnSearch.install = (Vue: Application) => {
+  Vue.component(VtmnSearch.name, VtmnSearch);
+};
+
+export default VtmnSearch;