+++ packages/sources/vue/src/components/overlays/VtmnToast/index.ts
@@ -1,0 +1,8 @@
+import { App as Application } from 'vue';
+import VtmnToast from './VtmnToast.vue';
+
+VtmnToast.install = (Vue: Application) => {
+  Vue.component(VtmnToast.name, VtmnToast);
+};
+
+export default VtmnToast;