+++ packages/sources/css/presets/tailwind/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.33.98](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.97...@vtmn/css-tailwind-preset@0.33.98) (2022-04-14)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.97](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.96...@vtmn/css-tailwind-preset@0.33.97) (2022-04-08)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.96](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.95...@vtmn/css-tailwind-preset@0.33.96) (2022-04-08)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.95](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.94...@vtmn/css-tailwind-preset@0.33.95) (2022-04-08)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.94](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.93...@vtmn/css-tailwind-preset@0.33.94) (2022-04-07)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.93](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.92...@vtmn/css-tailwind-preset@0.33.93) (2022-04-06)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.92](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.91...@vtmn/css-tailwind-preset@0.33.92) (2022-04-06)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.33.91](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.90...@vtmn/css-tailwind-preset@0.33.91) (2022-04-05)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
 ## [0.33.90](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.33.89...@vtmn/css-tailwind-preset@0.33.90) (2022-04-04)
 
 **Note:** Version bump only for package @vtmn/css-tailwind-preset